% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessing.R
\name{dimensionReduction}
\alias{dimensionReduction}
\title{Dimension reduction on data using PCAmethods}
\usage{
dimensionReduction(dat, method = c("svd", "nlpca"), cumR2 = 0.95,
  targetK = 20L)
}
\arguments{
\item{dat}{A data frame to process.}

\item{method}{A character string indicating the type of PCA: either \dQuote{svd}
or \dQuote{nlpca}.}

\item{cumR2}{A numeric value indicating the cumulative R2 value desired, used to
determine the number of dimensions to extract or use.}

\item{targetK}{An integer indicating the target number of variables desired.
When extracting raw variables, tries to (very) roughly match this target.}
}
\value{
A list.
}
\description{
This function uses PCA to reduce dimensionality by calculating
dimension scores and by listing the top variables for each
dimension.
}
\examples{
dimensionReduction(
 dat = mtcars,
 method = "svd",
 cumR2 = .95,
 targetK = 6)
dimensionReduction(
 dat = mtcars,
 method = "nlpca",
 cumR2 = .95,
 targetK = 6)
}
